# twit/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  DeletedAt:
    $ref: '#/definitions/NullTime'
  LoginData:
    properties:
      access-token:
        type: string
        x-go-name: AccessToken
      user:
        $ref: '#/definitions/User'
    type: object
    x-go-package: twit/models/responses
  LoginUserRequest:
    properties:
      email:
        type: string
        x-go-name: Email
      password:
        type: string
        x-go-name: Password
    type: object
    x-go-package: twit/models/requests
  LoginUserResponse:
    properties:
      data:
        $ref: '#/definitions/LoginData'
      message:
        type: string
        x-go-name: Message
      success:
        type: boolean
        x-go-name: Success
    type: object
    x-go-package: twit/models/responses
  NullTime:
    description: |-
      NullTime implements the Scanner interface so
      it can be used as a scan destination, similar to NullString.
    properties:
      Time:
        format: date-time
        type: string
      Valid:
        type: boolean
    title: NullTime represents a time.Time that may be null.
    type: object
    x-go-package: database/sql
  RegisterUserRequest:
    properties:
      email:
        type: string
        x-go-name: Email
      password:
        type: string
        x-go-name: Password
      username:
        type: string
        x-go-name: Username
    type: object
    x-go-package: twit/models/requests
  Response:
    properties:
      data:
        type: object
        x-go-name: Data
      message:
        type: string
        x-go-name: Message
      success:
        type: boolean
        x-go-name: Success
    type: object
    x-go-package: twit/models/responses
  User:
    properties:
      DeletedAt:
        $ref: '#/definitions/DeletedAt'
      createdAt:
        format: date-time
        type: string
        x-go-name: CreatedAt
      email:
        type: string
        x-go-name: Email
      id:
        format: uint64
        type: integer
        x-go-name: ID
      password:
        type: string
        x-go-name: Password
      updatedAt:
        format: date-time
        type: string
        x-go-name: UpdatedAt
      username:
        type: string
        x-go-name: Username
    type: object
    x-go-package: twit/models
host: localhost:9090
info:
  description: Documentation of our Twit Application.
  title: Twit Application.
  version: 1.0.0
paths:
  /auth/login:
    post:
      operationId: loginUser
      parameters:
      - description: This text will appear as description of your request body.
        in: body
        name: Body
        schema:
          $ref: '#/definitions/LoginUserRequest'
      responses:
        "200":
          $ref: '#/responses/loginResponse'
      summary: Login User.
      tags:
      - Authentication
  /auth/register:
    post:
      operationId: registerUser
      parameters:
      - description: This text will appear as description of your request body.
        in: body
        name: Body
        schema:
          $ref: '#/definitions/RegisterUserRequest'
      responses:
        "200":
          $ref: '#/responses/registerResponse'
      summary: Register User.
      tags:
      - Authentication
  /user/profile:
    get:
      operationId: getProfile
      responses:
        "200":
          $ref: '#/responses/getProfileResponse'
      security:
      - Bearer:
        - '[]'
      summary: Get Profile.
      tags:
      - Authentication
produces:
- application/json
responses:
  getProfileResponse:
    description: This text will appear as description of your response body.
    schema:
      $ref: '#/definitions/LoginUserResponse'
  loginResponse:
    description: This text will appear as description of your response body.
    schema:
      $ref: '#/definitions/LoginUserResponse'
  registerResponse:
    description: This text will appear as description of your response body.
    schema:
      $ref: '#/definitions/Response'
schemes:
- http
- https
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"